Add options for 10,20,50L/p/s and Optimize some code
